Validating access to images by calling ds-license
Only the iIPImageRequest method has implemented this by now. We need the
name/path of the image in the quests, since this is used to call
ds-license . The filtering in ds-license is done on the resource_id
field, and not the filter on id. ds-license 1.1 has a new method for
this call.

If no access generate a temporary image (bufferedImage).
Maybe throw a 403 (forbidden) instead? Must be decided later in the
process.