Extend Solr schema with 'text' field and enable More Like This + highlight support
The 'text' field is intended for full/body text: Content from a scanned book,
subtitles from af TV-recording, article text from newspapers.

More Like This is used for finding documents similar to the first hit that
matches a query. Typically the query will be for a specific document id,
ensuring a singular match.

This commit also adds termVectors to the text_general field for better
highlighting support. This works well in tandem with More Like This.