Add caching to DsLicenseClient
To handle the use case of checking for access to bitmap tiles (many concurrent
checks for the same ID with the same credentials), caching has been added to
the DsLicenseClient. It is enabled per default with max size 100 and max age 60
seconds, but as a precaution the old constructor has been deprecated, nudging
users to switch to the cache-explicit new constructors.

This also adds support for constructing the DsLicenseClient directly from the
project config, if the config conforms to the stated structure.